AI/requirement.txt
openai==1.108.0
python-dotenv==1.1.1
